backend/requirements.txt
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
azure-core==1.29.6
azure-storage-blob==12.19.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==5.0.1
django-storages==1.14.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
filelock==3.13.1
exceptiongroup==1.2.0
firebase-admin==6.3.0
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.7
packaging==23.2
networkx==3.2.1
numpy==1.26.3
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
sympy==1.12
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0